feat: implement new feature and update documentation refactor: improve code structure and fix minor issues docs: update README with project details and setup instructions chore: update dependencies and configuration files fix: resolve bug in [specific functionality]
