+++ src/database/enums/viewing-status.enum.ts
@@ -1,0 +1,6 @@
+export enum ViewingStatus {
+    PENDING = 'pending',
+    CONFIRMED = 'confirmed',
+    CANCELLED = 'cancelled',
+    COMPLETED = 'completed',
+}
