+++ src/database/schemas/property-viewing.schema.ts
@@ -1,25 +1,9 @@
 import { Prop, Schema, SchemaFactory } from '@nestjs/mongoose';
 import { Document, Types } from 'mongoose';
 import { Property } from './property.schema';
+import { ViewingStatus } from '../enums/viewing-status.enum';
+import { IPropertyViewing } from '../interfaces/property-viewing.interface';
 
-export enum ViewingStatus {
-  PENDING = 'pending',
-  CONFIRMED = 'confirmed',
-  CANCELLED = 'cancelled',
-  COMPLETED = 'completed',
-}
-
-export interface IPropertyViewing {
-  propertyId: Types.ObjectId;
-  userId: string;
-  agentId?: string;
-  startTime: Date;
-  endTime: Date;
-  status: ViewingStatus;
-  notes?: string;
-  createdAt: Date;
-  updatedAt: Date;
-}
 
 @Schema({ timestamps: true })
 export class PropertyViewing extends Document implements IPropertyViewing {
@@ -39,10 +23,10 @@ export class PropertyViewing extends Document implements IPropertyViewing {
   @Prop({ type: Date, required: true })
   endTime: Date;
 
-  @Prop({ 
-    type: String, 
+  @Prop({
+    type: String,
     enum: Object.values(ViewingStatus),
-    default: ViewingStatus.PENDING 
+    default: ViewingStatus.PENDING
   })
   status: ViewingStatus;
 
@@ -56,7 +40,7 @@ export class PropertyViewing extends Document implements IPropertyViewing {
 export type PropertyViewingDocument = PropertyViewing & Document;
 export const PropertyViewingSchema = SchemaFactory.createForClass(PropertyViewing);
 
-// Create compound index to prevent double bookings for the same property and time slot
+// Compound index to prevent double bookings and speed up queries
 PropertyViewingSchema.index(
   { propertyId: 1, startTime: 1, endTime: 1 },
   { unique: true, partialFilterExpression: { status: { $ne: ViewingStatus.CANCELLED } } }
