+++ src/property-viewings/interfaces/available-slots.interface.ts
@@ -1,0 +1,5 @@
+export interface AvailableSlot {
+    startTime: Date;
+    endTime: Date;
+    available: boolean;
+}
